+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Sprite3D` which works like regular PixiJS 2D sprites but in 3D space.
 - Added `fromHex` function to `Color`.
 
+### Changed
+- Renamed `morphWeights` to `targetWeights` for `Mesh3D`.
+
 ### Fixed
 - Fixed issues which caused `Camera`, `PostProcessingSprite` and `PickingHitArea` to not work correctly when renderer resolution was more than 1.